casting for subplot

+++ dataLoad.py
@@ -33,7 +33,7 @@ images = images.numpy()
 # plot the images in the batch, along with the corresponding labels
 fig = plt.figure(figsize=(25, 4))
 for idx in np.arange(batch_size):
-    ax = fig.add_subplot(2, batch_size/2, idx+1, xticks=[], yticks=[])
+    ax = fig.add_subplot(2, int(batch_size/2), idx+1, xticks=[], yticks=[])
     ax.imshow(np.squeeze(images[idx]), cmap='gray')
     # print out the correct label for each image
     # .item() gets the value contained in a Tensor
